Retry four times if Pfam download and processing fails

+++ scripts/plotting/pfam.py
@@ -6,7 +6,7 @@ import pandas as pd
 import subprocess
 
 from scripts import __logger_name__
-logger = daiquiri.getLogger(__logger_name__ + ".annotations.stability_change")
+logger = daiquiri.getLogger(__logger_name__ + ".annotations.pfam")
 
 logging.getLogger('urllib3.connectionpool').setLevel(logging.WARNING)
 
@@ -40,37 +40,47 @@ def get_pfam(seq_df, output_tsv):
     and Pfam ID to Transcript ID mapping.
     """
     
-    try:
-        # Pfam coordinates
-        logger.debug("Downloading and parsing Pfam coordinates...")
-        url_query = 'http://www.ensembl.org/biomart/martservice?query='
-        query = '<?xml version="1.0" encoding="UTF-8"?><!DOCTYPE Query><Query  virtualSchemaName = "default" formatter = "TSV" header = "0" uniqueRows = "0" count = "" datasetConfigVersion = "0.6" ><Dataset name = "hsapiens_gene_ensembl" interface = "default" ><Attribute name = "ensembl_gene_id" /><Attribute name = "ensembl_transcript_id" /><Attribute name = "pfam_start" /><Attribute name = "pfam_end" /><Attribute name = "pfam" /></Dataset></Query>'
-        url = url_query + query
-        command = [f"wget", "-q", "-O", "pfam_coordinates.tsv", url]
-        subprocess.run(command)
-        pfam = pd.read_csv("pfam_coordinates.tsv", sep="\t", header=None)
-        pfam.columns = ["Ens_Gene_ID", "Ens_Transcr_ID", "Pfam_start", "Pfam_end", "Pfam_ID"]
+    status = "INIT"
+    i = 0
+    
+    while status != "PASS":
+        if i < 5:
+            try:
+                # Pfam coordinates
+                logger.debug("Downloading and parsing Pfam coordinates...")
+                url_query = 'http://www.ensembl.org/biomart/martservice?query='
+                query = '<?xml version="1.0" encoding="UTF-8"?><!DOCTYPE Query><Query  virtualSchemaName = "default" formatter = "TSV" header = "0" uniqueRows = "0" count = "" datasetConfigVersion = "0.6" ><Dataset name = "hsapiens_gene_ensembl" interface = "default" ><Attribute name = "ensembl_gene_id" /><Attribute name = "ensembl_transcript_id" /><Attribute name = "pfam_start" /><Attribute name = "pfam_end" /><Attribute name = "pfam" /></Dataset></Query>'
+                url = url_query + query
+                command = [f"wget", "-q", "-O", "pfam_coordinates.tsv", url]
+                subprocess.run(command)
+                pfam = pd.read_csv("pfam_coordinates.tsv", sep="\t", header=None)
+                pfam.columns = ["Ens_Gene_ID", "Ens_Transcr_ID", "Pfam_start", "Pfam_end", "Pfam_ID"]
 
-        # ID database
-        logger.debug("Downloading and parsing Pfam ID database...")
-        url = "https://ftp.ebi.ac.uk/pub/databases/Pfam/current_release/database_files/pfamA.txt.gz"
-        command = ["wget", "-q", "-O", "pfam_id.tsv.gz", url]
-        subprocess.run(command)
-        pfam_id = pd.read_csv("pfam_id.tsv.gz", compression='gzip', sep='\t', header=None).iloc[:,[0,1,3]]
-        pfam_id.columns = "Pfam_ID", "Pfam_name", "Pfam_description"
+                # ID database
+                logger.debug("Downloading and parsing Pfam ID database...")
+                url = "https://ftp.ebi.ac.uk/pub/databases/Pfam/current_release/database_files/pfamA.txt.gz"
+                command = ["wget", "-q", "-O", "pfam_id.tsv.gz", url]
+                subprocess.run(command)
+                pfam_id = pd.read_csv("pfam_id.tsv.gz", compression='gzip', sep='\t', header=None).iloc[:,[0,1,3]]
+                pfam_id.columns = "Pfam_ID", "Pfam_name", "Pfam_description"
 
-        # Merge and save
-        pfam = pfam.merge(pfam_id, how="left", on="Pfam_ID")
-        pfam = pfam.dropna(how="all", subset=["Pfam_start", "Pfam_end"]).reset_index(drop=True)
-        pfam = add_pfam_metadata(pfam, seq_df)
-        pfam.to_csv(output_tsv, index=False, sep="\t")
-        
-        # Delete temp files
-        os.remove("pfam_coordinates.tsv")
-        os.remove("pfam_id.tsv.gz")
-        
-        return pfam
-        
-    except Exception as e:
-        logger.error('Download Pfam: FAIL')
-        logger.error(f"Error while downloading Pfam: {e}")
+                # Merge and save
+                pfam = pfam.merge(pfam_id, how="left", on="Pfam_ID")
+                pfam = pfam.dropna(how="all", subset=["Pfam_start", "Pfam_end"]).reset_index(drop=True)
+                pfam = add_pfam_metadata(pfam, seq_df)
+                pfam.to_csv(output_tsv, index=False, sep="\t")
+                
+                # Delete temp files
+                os.remove("pfam_coordinates.tsv")
+                os.remove("pfam_id.tsv.gz")
+                status = "PASS"
+                
+                return pfam
+                
+            except Exception as e:
+                status = "FAIL"
+                logger.warning(f"Error while downloading Pfam: {e}")
+                logger.warning("Retrying download...")
+                i += 1
+        else:
+            logger.error(f'Download Pfam: {status}')
